Add factory method for creating new class instances in CustomJS (#78)
This commit introduces a new factory method to the CustomJS evalFile function, allowing the creation of new instances of dynamically imported classes. This change enables users to instantiate new objects as needed, instead of relying solely on a single, shared instance. The method `createClassNameInstance` has been added to the `window.customJS` object, which returns a new instance of the specified class each time it is called.